Exercise/requirements.txt
numpy
pydicom
pandas
matplotlib
pillow